telegram_bot/requirements.txt
aiogram>=3.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
